# encoding/csv: optimize Write by giving fieldNeedsQuotes a fast path for when Comma is ascii